# openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW - Resources API
  description: API para gerenciamento de recursos, categorias e features
  version: 1.0.0
servers:
  - url: http://localhost:8187
    description: Servidor de desenvolvimento (externo)
  - url: http://resources:8080
    description: Servidor interno (Docker)

tags:
  - name: Categories
    description: Gerenciamento de categorias
  - name: Features
    description: Gerenciamento de features
  - name: Resources
    description: Gerenciamento de recursos
  - name: FeatureValues
    description: Gerenciamento de valores de features
  - name: ValueTypes
    description: Tipos de valores disponíveis

paths:
  /api/v1/categories:
    post:
      tags:
        - Categories
      summary: Criar uma nova categoria
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Categories
      summary: Listar todas as categorias
      operationId: findAllCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/categories/{categoryId}:
    get:
      tags:
        - Categories
      summary: Buscar categoria por ID
      operationId: findOneCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados da categoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Categories
      summary: Atualizar categoria
      operationId: updateCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Categoria atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Categories
      summary: Substituir categoria
      operationId: replaceCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Categoria substituída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Categories
      summary: Deletar categoria
      operationId: deleteCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Categoria deletada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/categories/{categoryId}/resources:
    get:
      tags:
        - Categories
      summary: Listar recursos de uma categoria
      operationId: getCategoryResources
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de recursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Categories
      summary: Criar recurso em uma categoria
      operationId: createCategoryResource
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Recurso criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/resources:
    post:
      tags:
        - Resources
      summary: Criar um novo recurso
      operationId: createResource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Recurso criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Resources
      summary: Listar todos os recursos
      operationId: findAllResources
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de recursos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/resources/{resourceId}:
    get:
      tags:
        - Resources
      summary: Buscar recurso por ID
      operationId: findOneResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dados do recurso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Resources
      summary: Atualizar recurso (parcial)
      operationId: updateResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceUpdateRequest'
      responses:
        '200':
          description: Recurso atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Resources
      summary: Substituir recurso (completo)
      operationId: replaceResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '200':
          description: Recurso substituído
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Resources
      summary: Deletar recurso
      operationId: deleteResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recurso deletado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/features:
    post:
      tags:
        - Features
      summary: Criar uma nova feature
      operationId: createFeature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureCreateRequest'
      responses:
        '201':
          description: Feature criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Features
      summary: Listar todas as features
      operationId: findAllFeatures
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de features
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/features/{featureId}:
    get:
      tags:
        - Features
      summary: Buscar feature por ID
      operationId: findOneFeature
      security:
        - bearerAuth: []
      parameters:
        - name: featureId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados da feature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Features
      summary: Atualizar feature (parcial)
      operationId: updateFeature
      security:
        - bearerAuth: []
      parameters:
        - name: featureId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureUpdateRequest'
      responses:
        '200':
          description: Feature atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Features
      summary: Substituir feature (completo)
      operationId: replaceFeature
      security:
        - bearerAuth: []
      parameters:
        - name: featureId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureUpdateRequest'
      responses:
        '200':
          description: Feature substituída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Features
      summary: Deletar feature
      operationId: deleteFeature
      security:
        - bearerAuth: []
      parameters:
        - name: featureId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Feature deletada
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/features/categories/{categoryId}:
    get:
      tags:
        - Features
      summary: Buscar features por categoria
      operationId: getFeaturesByCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de features da categoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/feature-values:
    post:
      tags:
        - FeatureValues
      summary: Criar um novo feature value
      operationId: createFeatureValue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueCreateRequest'
      responses:
        '201':
          description: Feature value criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - FeatureValues
      summary: Listar todos os feature values
      operationId: findAllFeatureValues
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/feature-values/{featureValueId}:
    get:
      tags:
        - FeatureValues
      summary: Buscar feature value por ID
      operationId: findOneFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados do feature value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - FeatureValues
      summary: Atualizar feature value (parcial)
      operationId: updateFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueUpdateRequest'
      responses:
        '200':
          description: Feature value atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - FeatureValues
      summary: Deletar feature value
      operationId: deleteFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Feature value deletado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/feature-values/resource/{resourceId}:
    get:
      tags:
        - FeatureValues
      summary: Buscar feature values por recurso
      operationId: getFeatureValuesByResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de feature values do recurso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/feature-values/feature/{featureId}:
    get:
      tags:
        - FeatureValues
      summary: Buscar feature values por feature
      operationId: getFeatureValuesByFeature
      security:
        - bearerAuth: []
      parameters:
        - name: featureId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de feature values da feature
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/feature-values/resources/{resourceId}/features:
    post:
      tags:
        - FeatureValues
      summary: Criar feature value para um recurso
      operationId: createFeatureValueForResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueCreateRequest'
      responses:
        '201':
          description: Feature value criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags:
        - FeatureValues
      summary: Listar feature values de um recurso
      operationId: getFeatureValuesForResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/feature-values/resources/{resourceId}/features/{featureValueId}:
    get:
      tags:
        - FeatureValues
      summary: Buscar feature value específico
      operationId: getFeatureValueForResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados do feature value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - FeatureValues
      summary: Atualizar feature value
      operationId: updateFeatureValueForResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueUpdateRequest'
      responses:
        '200':
          description: Feature value atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - FeatureValues
      summary: Deletar feature value
      operationId: deleteFeatureValueForResource
      security:
        - bearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
        - name: featureValueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Feature value deletado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/value-types:
    get:
      tags:
        - ValueTypes
      summary: Listar tipos de valores disponíveis (enum)
      operationId: getValueTypes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tipos de valores
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/health:
    get:
      tags:
        - Health
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Serviço está funcionando

  /api/v1/metrics:
    get:
      tags:
        - Metrics
      summary: Métricas Prometheus
      operationId: getMetrics
      responses:
        '200':
          description: Métricas em formato Prometheus
          content:
            text/plain:
              schema:
                type: string

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    CategoryCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    CategoryUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    CategoryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string

    ResourceCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    ResourceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    ResourceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    FeatureCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    FeatureUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    FeatureResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    FeatureValueCreateRequest:
      type: object
      properties:
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid

    FeatureValueUpdateRequest:
      type: object
      properties:
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid

    FeatureValueResponse:
      type: object
      properties:
        id:
          type: string
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid

    Error:
      type: object
      properties:
        detail:
          type: string