# src/controllers/resource.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Put,
  Param,
  Query,
  Delete,
  ValidationPipe,
  UseGuards,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiParam, ApiBearerAuth } from '@nestjs/swagger';
import { ResourceService } from '../services/resource.service';
import { CreateResourceDto } from '../dto/create-resource.dto';
import { UpdateResourceDto } from '../dto/update-resource.dto';
import { Resource } from '../entities/resource.entity';
import { AuthGuard } from '../guards/auth.guard';

@ApiTags('resources')
@ApiBearerAuth()
@UseGuards(AuthGuard)
@Controller('resources')
export class ResourceController {
  constructor(private readonly resourceService: ResourceService) { }

  @Post()
  @ApiOperation({ summary: 'Create a new resource' })
  @ApiResponse({
    status: 201,
    description: 'The resource has been successfully created.',
    type: Resource,
  })
  async create(
    @Body(ValidationPipe) createResourceDto: CreateResourceDto,
  ): Promise<Resource> {
    return await this.resourceService.create(createResourceDto);
  }

  @Get()
  @ApiOperation({ summary: 'Get all resources' })
  @ApiResponse({
    status: 200,
    description: 'Return all resources.',
    type: [Resource],
  })
  async findAll(@Query() query?: any): Promise<Resource[]> {
    // if query contains categoryId, delegate to findByCategory
    if (query && query.categoryId) {
      return await this.resourceService.findByCategory(query.categoryId);
    }
    return await this.resourceService.findAll();
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get a resource by id' })
  @ApiParam({ name: 'id', description: 'Resource ID', type: 'string' })
  @ApiResponse({
    status: 200,
    description: 'Return the resource.',
    type: Resource,
  })
  @ApiResponse({ status: 404, description: 'Resource not found.' })
  async findOne(@Param('id') id: string): Promise<Resource> {
    return await this.resourceService.findOne(id);
  }

  @Patch(':id')
  @ApiOperation({ summary: 'Update a resource' })
  @ApiParam({ name: 'id', description: 'Resource ID', type: 'string' })
  @ApiResponse({
    status: 200,
    description: 'The resource has been successfully updated.',
    type: Resource,
  })
  @ApiResponse({ status: 404, description: 'Resource not found.' })
  async update(
    @Param('id') id: string,
    @Body(ValidationPipe) updateResourceDto: UpdateResourceDto,
  ): Promise<Resource> {
    return await this.resourceService.update(id, updateResourceDto);
  }

  @Put(':id')
  @ApiOperation({ summary: 'Replace a resource' })
  @ApiParam({ name: 'id', description: 'Resource ID', type: 'string' })
  async replace(
    @Param('id') id: string,
    @Body(ValidationPipe) updateResourceDto: UpdateResourceDto,
  ): Promise<Resource> {
    return await this.resourceService.update(id, updateResourceDto);
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Delete a resource' })
  @ApiParam({ name: 'id', description: 'Resource ID', type: 'string' })
  @ApiResponse({
    status: 200,
    description: 'The resource has been successfully deleted.',
  })
  @ApiResponse({ status: 404, description: 'Resource not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    return await this.resourceService.remove(id);
  }
}